consolidating utils into utils.js
